Extract GrpcKademliaRPCServer from GrpcKademliaRPC
